+++ DeusCore/PacketTest.cpp
@@ -14,7 +14,7 @@ namespace DeusCore
 	void PacketTest::OnDeserialize(Buffer512 & buffer)
 	{
 		// get the size of the string
-		size_t dataSize;
+		uint32_t dataSize;
 		DeserializeData(buffer, dataSize);
 
 		// get the string
@@ -27,7 +27,7 @@ namespace DeusCore
 
 	void PacketTest::OnSerialize(Buffer512 & buffer) const
 	{
-		size_t dataSize = m_message.size() + 1; // +1 to add the \0 of string
+		uint32_t dataSize = m_message.size() + 1; // +1 to add the \0 of string
 		SerializeData(buffer, dataSize); // SerializeData only for primitive
 
 		//  then we add the string
